move cve_id_comparator to helpers for the general use
